# backend/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RBAC 权限管理系统 API",
    "version": "1.0.0",
    "description": "RBAC权限管理系统后端接口文档",
    "contact": {
      "name": "技术支持",
      "email": "support@example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "开发环境"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/users/{id}/roles": {
      "put": {
        "summary": "分配用户角色",
        "tags": [
          "用户管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "number"
            },
            "description": "用户ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "roleIds": {
                    "type": "array",
                    "items": {
                      "type": "number"
                    },
                    "description": "角色ID列表"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "分配成功"
          }
        }
      }
    },
...
    "/roles": {
      "get": {
        "summary": "获取角色列表",
        "tags": [
          "角色管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取角色列表"
          }
        }
      },
      "post": {
        "summary": "新增角色",
        "tags": [
          "角色管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "code",
                  "name"
                ],
                "properties": {
                  "code": {
                    "type": "string",
                    "description": "角色编码"
                  },
                  "name": {
                    "type": "string",
                    "description": "角色名称"
                  },
                  "description": {
                    "type": "string"
                  },
                  "sort": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功"
          }
        }
      }
    },
    "/roles/{id}": {
      "get": {
        "summary": "获取角色详情",
        "tags": [
          "角色管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "number"
            },
            "description": "角色ID"
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取角色详情"
          }
        }
      },
      "put": {
        "summary": "编辑角色",
        "tags": [
          "角色管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "number"
            },
            "description": "角色ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "sort": {
                    "type": "number"
                  },
                  "status": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "编辑成功"
          }
        }
      },
      "delete": {
        "summary": "删除角色",
        "tags": [
          "角色管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "number"
            },
            "description": "角色ID"
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          }
        }
      }
    },
    "/roles/{id}/menus": {
      "get": {
        "summary": "获取角色菜单",
        "tags": [
          "角色管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "number"
            },
            "description": "角色ID"
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取角色菜单ID列表"
          }
        }
      },
      "post": {
        "summary": "分配角色权限",
        "tags": [
          "角色管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "number"
            },
            "description": "角色ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "menuIds": {
                    "type": "array",
                    "items": {
                      "type": "number"
                    },
                    "description": "菜单ID列表"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "权限分配成功"
          }
        }
      }
    },
    "/menus/tree": {
      "get": {
        "summary": "获取菜单树",
        "tags": [
          "菜单管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取菜单树",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number"
                          },
                          "parentId": {
                            "type": "number"
                          },
                          "name": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "path": {
                            "type": "string"
                          },
                          "children": {
                            "type": "array",
                            "items": {
                              "type": "object"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/menus/routes": {
      "get": {
        "summary": "获取用户路由菜单",
        "tags": [
          "菜单管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "获取当前登录用户可访问的路由菜单（用于前端动态路由）",
        "responses": {
          "200": {
            "description": "成功获取用户路由"
          }
        }
      }
    },
    "/menus": {
      "get": {
        "summary": "获取菜单列表",
        "tags": [
          "菜单管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取菜单列表"
          }
        }
      },
      "post": {
        "summary": "新增菜单",
        "tags": [
          "菜单管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "type",
                  "title"
                ],
                "properties": {
                  "parentId": {
                    "type": "number",
                    "description": "父菜单ID"
                  },
                  "path": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "icon": {
                    "type": "string"
                  },
                  "component": {
                    "type": "string"
                  },
                  "sort": {
                    "type": "number"
                  },
                  "type": {
                    "type": "number",
                    "description": "0目录 1菜单 2按钮"
                  },
                  "title": {
                    "type": "string"
                  },
                  "permission": {
                    "type": "string"
                  },
                  "hidden": {
                    "type": "number"
                  },
                  "status": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功"
          }
        }
      }
    },
    "/menus/{id}": {
      "get": {
        "summary": "获取菜单详情",
        "tags": [
          "菜单管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "number"
            },
            "description": "菜单ID"
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取菜单详情"
          }
        }
      },
      "put": {
        "summary": "编辑菜单",
        "tags": [
          "菜单管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "number"
            },
            "description": "菜单ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "parentId": {
                    "type": "number"
                  },
                  "path": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "icon": {
                    "type": "string"
                  },
                  "component": {
                    "type": "string"
                  },
                  "sort": {
                    "type": "number"
                  },
                  "title": {
                    "type": "string"
                  },
                  "permission": {
                    "type": "string"
                  },
                  "status": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "编辑成功"
          }
        }
      },
      "delete": {
        "summary": "删除菜单",
        "tags": [
          "菜单管理"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "number"
            },
            "description": "菜单ID"
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          },
          "400": {
            "description": "存在子菜单无法删除"
          }
        }
      }
    }
  },
...
}